+++ jano_core/src/main/java/ec/com/uce/jano/entities/Recaudacion.java
@@ -1,9 +1,17 @@
 package ec.com.uce.jano.entities;
 
 import java.io.Serializable;
-import javax.persistence.*;
-import java.math.BigDecimal;
 import java.util.Date;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Temporal;
+import javax.persistence.TemporalType;
 
 
 /**
@@ -33,7 +41,7 @@ public class Recaudacion implements Serializable {
 	private String peridoRecaudacion;
 
 	@Column(name="valor_recaudacion")
-	private BigDecimal valorRecaudacion;
+	private double valorRecaudacion;
 
 	//bi-directional many-to-one association to Afectacion
 	@ManyToOne
@@ -96,11 +104,11 @@ public class Recaudacion implements Serializable {
 		this.peridoRecaudacion = peridoRecaudacion;
 	}
 
-	public BigDecimal getValorRecaudacion() {
+	public double getValorRecaudacion() {
 		return this.valorRecaudacion;
 	}
 
-	public void setValorRecaudacion(BigDecimal valorRecaudacion) {
+	public void setValorRecaudacion(double valorRecaudacion) {
 		this.valorRecaudacion = valorRecaudacion;
 	}
